Homework 07 finished
