+++ src/App.tsx
@@ -2,7 +2,7 @@ import type { JSX } from "react";
 import "./App.css";
 // lessons
 // import Lesson_06 from './lessons/Lesson06/Lesson_06'
-import Lesson_07 from "./lessons/Lesson_07/Lesson_07";
+// import Lesson_07 from "./lessons/Lesson_07/Lesson_07";
 //homeworks
 // import Homework_06 from './homeworks/Homeworks_06/Homework_06'
 import Homework_07 from "homeworks/Homework_07/Homework_07";
