+++ src/components/Feedback/Feedback.tsx
@@ -1,37 +1,9 @@
-import { useState } from "react";
-
 import Button from "components/Button/Button";
-
+import {type  FeedbackProps } from "./types";
 import "./styles.css";
 
-function Feedback() {
-  const [like, setLike] = useState(0); // Если мы ничего не передаем в качестве аргумента функции useState(), то 1 элемент массива, который эта функция возвращает === undefined
-  const [dislike, setDislike] = useState(0);
-
-  const onLike = () => {
-    setLike((prevValue) => {
-      console.log(prevValue);
-      return prevValue + 1;
-    });
-    setLike((prevValue) => {
-      console.log(prevValue);
-      return prevValue + 1;
-    });
-    setLike((prevValue) => {
-      console.log(prevValue);
-      return prevValue + 1;
-    });
-  };
-
-  const onDislike = () => {
-    setDislike((prevValue) => prevValue + 1);
-  };
-
-  const resetResults = () => {
-    setLike(0);
-    setDislike(0);
-  };
-
+function Feedback({like, dislike, onLike, onDislike, resetResults}:FeedbackProps) {
+ 
   return (
     <div className="feedback-wrapper">
       <div className="feedback-control">
