+++ src/components/Feedback/types.ts
@@ -1,0 +1,7 @@
+export interface FeedbackProps {
+  like: number;
+  dislike: number;
+  onLike: () => void;
+  onDislike: () => void;
+  resetResults: () => void;
+}
